pyproject.toml
[tool.poetry]
name = "sonara"
version = "0.1.0"
description = "A Python project for voice recognition and translation."
authors = ["Wesley Guo <yuxing.g@gmail.com>"]
readme = "README.md"
packages = [{ include = "sonara" }]

[tool.poetry.dependencies]
python = "^3.10"
websockets = "15.0"
azure-cognitiveservices-speech = "^1.42.0"
python-dotenv = "^1.0.1"
ollama = "0.4.7"
groq = "0.18.0"

[tool.poetry.group.dev.dependencies]
pytest = "8.3.4"
pytest-cov = "6.0.0"
black = "25.1.0"
isort = "6.0.0"
pytest-asyncio = "^0.25.3"

[tool.poetry.scripts]
server = "sonara.server:main_entrypoint"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest]
log_cli = true
log_cli_level = "DEBUG"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::RuntimeWarning:unittest.mock:2188",
    "ignore::RuntimeWarning:unittest.mock:767"
]
asyncio_default_fixture_loop_scope = "function"
